backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.6
gunicorn==23.0.0
pytest==8.3.3
black==24.10.0
flake8==7.1.1
requests==2.32.4
openpyxl==3.1.2